add remoteMediator2

+++ app/src/main/java/com/example/rickandmorty/data/model/PageKey.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "pageKey")
 data class PageKey(
-    @PrimaryKey val id: Int?,
+    @PrimaryKey val id: Long?,
     val nextPageUrl: String?
 )
